音量持久化
- player_controller.h—添加QSettings头文件和成员变量
- player_controller.cpp—构造时读取上次音量setVolume时保存；playFile 重新打开SDL后恢复音量

+++ src/lyrics_parser.h
@@ -37,7 +37,7 @@ public:
     QString filePath() const { return filePath_; }
     void setFilePath(const QString &path);
 
-    int lineCount() const { return lines_.size(); }
+    int lineCount() const { return static_cast<int>(lines_.size()); }
 
     Q_INVOKABLE bool load(const QString &path);
     Q_INVOKABLE bool loadFromString(const QString &lrcContent);
