+++ src/search_result_model.cpp
@@ -24,13 +24,13 @@ void SearchResultModel::clear()
 
 int SearchResultModel::count() const
 {
-    return results_.size();
+    return static_cast<int>(results_.size());
 }
 
 int SearchResultModel::rowCount(const QModelIndex &parent) const
 {
     Q_UNUSED(parent);
-    return results_.size();
+    return static_cast<int>(results_.size());
 }
 
 // 根据角色返回对应的数据
@@ -40,27 +40,32 @@ QVariant SearchResultModel::data(const QModelIndex &index, int role) const
 
     const SearchResult &r = results_[index.row()];
     switch (role) {
-    case SongIdRole:    return r.songId;
-    case NameRole:      return r.name;
-    case ArtistRole:    return r.artist;
-    case AlbumRole:     return r.album;
-    case DurationRole:  return r.duration;
-    case CoverUrlRole:  return r.coverUrl;
-    default:            return {};
+    case SongIdRole:
+        return r.songId;
+    case NameRole:
+        return r.name;
+    case ArtistRole:
+        return r.artist;
+    case AlbumRole:
+        return r.album;
+    case DurationRole:
+        return r.duration;
+    case CoverUrlRole:
+        return r.coverUrl;
+    default:
+        return {};
     }
 }
 
 // 注册QML可用的角色名称
 QHash<int, QByteArray> SearchResultModel::roleNames() const
 {
-    return {
-        {SongIdRole,    "songId"},
-        {NameRole,      "name"},
-        {ArtistRole,    "artist"},
-        {AlbumRole,     "album"},
-        {DurationRole,  "duration"},
-        {CoverUrlRole,  "coverUrl"}
-    };
+    return {{SongIdRole, "songId"},
+            {NameRole, "name"},
+            {ArtistRole, "artist"},
+            {AlbumRole, "album"},
+            {DurationRole, "duration"},
+            {CoverUrlRole, "coverUrl"}};
 }
 
 int SearchResultModel::songId(int row) const
